# src/main/resources/openapi/order.yaml
openapi: 3.0.0
info:
  title: Order Management API
  version: 1.0.0
  description: API for updating orders.


#------valeena------------------------
paths:
  /users/orders/{trackingid}:
    parameters:
      - in: path
        name: trackingid
        required: true
        schema:
          type: string
        description: The tracking ID of the order to be updated
    patch:
      tags:
        - orders
      summary: Update Order
      description: Update the details of a specific order.
      requestBody:
        description: Order status object to update.
        content:
          application/json:
            schema:
              type: object
              properties:
                order_status:
                  type: string
                  description: The new status of the order.
              required:
                - order_status
      responses:
        '200':
          description: Order updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_status:
                    type: string
                    description: The status of the order.
                  tracking_id:
                    type: string
                    description: The tracking ID of the order.
                  message:
                    type: string
                    description: Success message.
                    example: Order updated successfully.
        '404':
          description: Order not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/orders/:
    parameters:
      - in: query
        name: page
        required: true
        schema:
          type: integer
        description: Page number to be viewed
      - in: query
        name: size
        required: true
        schema:
          type: integer
        description: Number of items per page
      - in: query
        name: status
        required: false
        schema:
          type: string
        description: Status of order
    get:
      tags:
        - orders
      summary: Get Orders
      description: Retrieve a list of orders for a user along with the total count of all orders and active orders.
      responses:
        '200':
          description: A list of orders with counts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_orders:
                    type: integer
                    description: The total count of all orders.
                  active_orders:
                    type: integer
                    description: The count of active orders.
                  orders:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: The unique identifier for the order.
                        order_status_id:
                          type: integer
                          description: The status identifier of the order.
                        user_id:
                          type: integer
                          description: The user who placed the order.
                        amount:
                          type: integer
                          description: The total amount for the order.
                        tracking_id:
                          type: string
                          description: Unique tracking identifier for the order.
                        created_at:
                          type: string
                          format: date-time
                          description: The timestamp when the order was created.
                        updated_at:
                          type: string
                          format: date-time
                          description: The timestamp when the order was last updated.
                        orderItems:
                          type: array
                          description: List of items in the order.
                          items:
                            $ref: '#/components/schemas/OrderItem'


  #------hassan------------------------
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: The error code or type.
        message:
          type: string
          description: A brief message describing the error.
        details:
          type: string
          description: Additional details about the error.
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the order item.
        quantity:
          type: integer
          description: The quantity of the order item.
        order_id:
          type: integer
          description: The identifier of the associated order.
        recipe_id:
          type: integer
          description: The identifier of the associated recipe.
        created_at:
          type: string
          format: date-time
          description: The timestamp when the order item was created.
        updated_at:
          type: string
          format: date-time
          description: The timestamp when the order item was last updated.

    #------------------hassan
...